test(stella-tui): the AGENTS and TRACES keymap rows get a file named for them
#4368 wants a guard that every `keymap::BINDINGS` row names a witness. The
blocker it hit was not that the AGENTS and TRACES rows are untested — it is
that their tests live in files whose names say nothing about those tabs, so
neither a guard nor a reader can find them.

Read against main, the coverage is: `⏎`, `a`, `n`, `x x` and `v` are pressed
through `handle_deck_key` in `deck_ui/tests/graph.rs` (which is otherwise the
GRAPH file picker), and the TRACES `f` cycle is pressed in
`deck_ui/tests/tabs.rs` (which is otherwise the tab bar). The issue's evidence
says none of those files presses an AGENTS or TRACES verb; that is not what
the files contain, and re-asserting behaviour that is already asserted would
have been a second copy rather than a witness.

So this moves them rather than rewriting them — the assertions are unchanged,
byte for byte — into `deck_ui/tests/agents.rs` and `deck_ui/tests/traces.rs`,
and adds the three rows that genuinely had no key press behind them:

* `r` (the second half of the `v / r` row) sends `AgentsRefresh` and shows the
  working state. The only other `AgentsRefresh` in the suite is the one
  `/agents` sends on opening the tab, which says nothing about the key.
* `r` and `f` are letters, so each gets the composer gate: with a prompt half
  typed they are the next character of it, never the verb.
* `f` with no agents registered stays unfiltered, rather than selecting an
  absent agent and rendering an empty tab.

Both new witnesses were run against a tree with their handler arm deleted:
`installed_r_reloads_the_list_from_disk` fails `Handled != Send(AgentsRefresh)`
and `traces_filter_cycles_through_agents_and_back` fails `None != Some("a")`.

Moved tests (none deleted): `slash_agents_opens_the_tab_on_the_installed_pane`,
`agents_list_ingest_updates_the_panel_out_of_band_and_clamps`,
`installed_enter_opens_the_editor_and_ctrl_s_saves_a_new_version`,
`editor_esc_discards_without_sending_and_typing_never_leaks`,
`create_flow_describes_picks_scope_and_dispatches_the_llm_draft`,
`agent_creating_dialog_survives_a_queued_snapshot_and_ends_in_the_detail_view`,
`agent_create_done_view_scrolls_like_the_skills_preview`,
`agent_failed_create_shows_the_error_in_the_dialog`,
`agent_creating_dialog_esc_hides_but_the_op_stays_busy`,
`create_flow_requires_a_description_and_esc_steps_back`,
`version_picker_pins_an_older_version_without_editing`,
`version_picker_on_the_already_pinned_version_sends_nothing`,
`installed_browse_letter_verbs_type_when_the_composer_has_text`,
`installed_x_twice_deletes_and_a_assumes` (graph.rs → agents.rs) and
`traces_filter_cycles_through_agents_and_back` (tabs.rs → traces.rs).

Closes #4429
